test_v2/pyproject.toml
[project]
name = "past-api-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.73.1",
    "google-generativeai>=0.8.6",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
